design fixes, car facts, fuelAPI, Fuel cards, reminder content

+++ src/sections/SectionProflile/SectionProfile.tsx
@@ -11,6 +11,7 @@ import {
   ProfileButtons,
   ProfileContainer,
   ProfileIcons,
+  ReminderContent,
   Section3Background,
   SectionContainer,
   SectionWrapper,
@@ -18,15 +19,19 @@ import {
 
 const SectionProfile = () => {
   const user = useAppSelector(authSliceSelectors.selectCurrentUser);
-  
+
   const dispatch = useAppDispatch();
 
-
   function logoutHandler() {
-     dispatch(usersSliceActions.logoutUser())
+    dispatch(usersSliceActions.logoutUser());
   }
   function deleteHandler() {
-  dispatch(authSliceActions.deleteCurrentUser())
+    const isConfirmed = window.confirm("Are you sure you want to delete your account? This action cannot be undone.");
+    if (isConfirmed) {
+      dispatch(authSliceActions.deleteCurrentUser())
+    } else {
+      console.log('Account deletion aborted.');
+    }
   }
 
   return (
@@ -36,13 +41,20 @@ const SectionProfile = () => {
           <Section3Background>
             <ProfileBackground>
               <ProfileButtons>
+                <ProfileIcons src={DeleteUser} onClick={deleteHandler} />
                 <ProfileIcons src={Logout} onClick={logoutHandler} />
-                <ProfileIcons src={DeleteUser} onClick={deleteHandler}/>
               </ProfileButtons>
               <ProfileContainer>
                 <FuelCard />
               </ProfileContainer>
-              <ProfileContainer>Здесь контет напоминания</ProfileContainer>
+              <ProfileContainer>
+                <ReminderContent>
+                Dear [Name], <br/>As seasons change, remember to switch your car's tires. German law requires winter tires during icy, snowy, and slushy
+                conditions, typically from autumn to spring. Summer tires are best when temperatures consistently exceed 7°C. Ensure your tires have
+                the correct tread depth: at least 1.6 mm by law, but ideally 3 mm for summer tires and 4 mm for winter tires for optimal safety. Stay
+                safe and enjoy your drive! Best regards, [Your Name/Your Company]
+              </ReminderContent>  
+              </ProfileContainer>
               <ProfileContainer>
                 <UserInfo />
               </ProfileContainer>
@@ -50,7 +62,7 @@ const SectionProfile = () => {
           </Section3Background>
         </SectionContainer>
       )}
-      <Footer>
+      <Footer>      
         <FooterIcons src={GitHub} />
         <FooterIcons src={Mail} />
       </Footer>
